requirements.txt
python-dotenv
fastapi
mysql-connector-python
uvicorn
pandas